Final version/requirements.txt
streamlit==1.32.0
dspy-ai==2.4.12
langchain==0.1.11
langchain-community==0.0.25
langchain-groq==0.1.4
pdfplumber==0.10.4
python-docx==1.1.0
beautifulsoup4==4.12.3
faiss-cpu==1.8.0
groq==0.4.1
google-search-results==2.4.2
google==3.0.0
huggingface-hub==0.23.0
numpy==1.26.4
requests==2.31.0